Edits all over the place, some performing better, some worse!
Made a short random dataset with 5,000 rows. Added n_short = 5000, and put n_short in all places we were putting rows, so then if we change the subset on the dataset, that changes automatically.

Gcomp with SL is working, although I don't know what the numbers mean.

SL is still not working. If it runs I get an error (missing data is currently not supported) but most times my computer just crashes.

I think the ITPW is working, although the numbers are kinda weird.

The bootstrap is a nightmare. I don't know what to do anymore. Now Shalika says to convert the outcome to binary. I tried this but code still not running.